dbDataType's first argument should be DBIDriver
With this I also switched to calling dbDataType directly instead of using db_data_type. Using real objects is rather tricky with the 'fields' argument for that function.
It's actually quite easy to misuse since list() is a valid argument to dbDataType for the PrestoDriver.
So we might want to disable/remove db_data_type until there is a more significant need.

Note that this PR stacks on top of #26.