Add subscripting support for arrays and maps
Note that these are sort of difficult to use since they need to be
wrapped with `remote()` for most calls.

There is also `getElement` that I could add to the mix that could be
used directly.

On the other hand, I don't want to over-implement.
We will need a separate subscripting operator when we add support for
rows, and I was thinking maybe we could use `getElement` for that
specifically.

Another alternative is to use `[` for arrays and maps, and
use `[[` for rows. But then we sort of break the parallel between maps
and lists.

Any advice?

Test Plan:

    library('dplyr')
    library('testthat')
    devtools::load_all('~/local/repositories/RPresto')
    devtools::test('~/local/repositories/RPresto', 'src_translate_env')