Handle non-standard numeric values
Specifically we need to special case the Inf and NaN values coming from Presto
as they are returned as strings. This causes R to fail if you have dplyr
installed because `bind_rows` will fail if columns have different types in the
data frames it is trying to combine. Even if you don't have dplyr installed this
will cause your numeric column to become a character because of R's implicit
conversion.

This is rebased on #39 to make the tests pass, we should merge that first.