Support for proper session time zone handling
Presto allows the setting of the timezone through the X-Presto-Time-Zone
header, we add proper support for that.

The session timezone is mainly used when:
- parsing strings as timestamps
- Timestamp representations of hive 'TIMESTAMP' columns (which are
  internally stored as unix times)
- Interval calculations on timestamp fields

See the discussion in #25.

My current approach is to automatically infer the time zone from the R session when available.
If that's not available through Sys.timezone() (which will unfortunately be the case most of the time), we give a warning and let presto assign its default.

I'm not sure about this behavior, since I imagine the warnings could be rather annoying.
Another alternative is to make the default 'UTC', clearly document this, and only change it if explicitly specified during dbConnect().
@cosinequanon, any thoughts?

@dain, could you check tests/testthat/test-session.timezone.R and
verify that my 'expected values' for the test cases are correct?