Fix locale and encoding handling
As #25 identified, some tests were failing when locale was not en_US.utf-8.

The issues were mostly around encodings of strings in the test cases.

One bug I identified was the use of toupper. With a Turkish locale, the uppercase of 'i' is 'İ' for instance.
To work around, we use 'stringi', which is already a dependency of 'httr' via 'stringr'.

Fixes #25.

For testing:

    Sys.setlocale('LC_CTYPE', 'C')
    library('dplyr'); library('DBI'); devtools::load_all('~/local/repositories/RPresto'); devtools::test('~/local/repositories/RPresto')

    Sys.setlocale('LC_CTYPE', 'en_US.utf-8')
    library('dplyr'); library('DBI'); devtools::load_all('~/local/repositories/RPresto'); devtools::test('~/local/repositories/RPresto')