Add warning for integer columns being corrupted by lacking 64-bit integer support
Presto BIGINT columns support 64-bit integers, whereas R is still
lacking full support for them. R silently casts any such value to a
double, which leads to loss of precision and is impossible to recover
afterwards.

jsonlite solves this by utilizing the bit64 package. That approach has
lots of caveats though, so we need to investigate more to see if we can
do the same.