Skip test cases where we need a test locale if we cannot set it
In order for `with_locale` in the tests to work, we need to know the OS-specific name
for the test locale. We do not have access to the Solaris test
environment that CRAN tests on, which results in the build failing for
that platform. So, if we cannot set the locale we just skip the
corresponding test case.