Add a db_query_fields implementation
dplyr is switching the default implementation of db_query_fields
from using `dbListFields` to using `dbFetch(result, 0)`
which does not work for Presto. So we use the old implementation
instead, with the deprecated `build_sql` approach. Once we start
depending on the latest version, we should move to the more canonical
approach.