use native class check for .null.to.na
`.null.to.na` takes majority of the time spent in `.extract.data` in a presto query. By using native `is.null` and `is.list` checks can speed up the function up to 2x. Raw typed column checking was the next hotspot in the function. By moving it outside, we gain additional non-negligible speed up.