Handle Decimal type as character
Storing the decimal type as a number doesn't actually make sense since that type
more closely matches a double. Furthermore Presto encodes the decimal type in
the JSON response as a string so in order to keep the arbitrary precision of the
type we need to store is a character. Users can then use their favorite
arbitrary precision library to convert the string into that class.

Fixes #60