Wrap the `sql` argument to `db_query_fields` with a `sql_subquery`
There is some behavioral change between dplyr 0.4.3 and 0.5.0
in the `sql` argument to `db_query_fields`. If it is a subquery (as
opposed to a table name), it is no longer wrapped in parentheses. We
then end up with an invalid SQL string that looks like:

    SELECT * FROM SELECT ...

Wrapping it with `sql_subquery` fixes the issue for 0.5.0. It still
works for 0.4.3, since it just adds another pair of parentheses.