Do not assume the first response will return column information
With recent changes to presto, the first response might not always
return the column information for a query.
This breaks the assumptions for `dbListFields(result)`.

We could:
- Advance the 'cursor' until we get a response with column information
- Reissue another query with `WHERE 1 = 0` which is hopefully light
  enough to return quickly.

The former might break some other assumptions. So I went for the latter
option. This required carrying the connection object alongside the
result object, which is not necessarily a bad thing, especially if we
implement things like `dbColumnInfo` in the future.