Lazily attach s3 methods
these methods don't exist unless dplyr is attached, in 3.6 there's a way
to lazily attach them. gate the namespace attachment behind 3.6.
unfortunately this means we have to manually edit the NAMESPACE file but
this is the easiest way forward right now, hopefully roxygen will
support things like this natively in the future