Use the delayed registration mechanism in R-devel for dplyr method implementations
Due to desire on our part to minimize dependencies, we try not to add
dplyr to Imports but keep it in Suggests. This means roxygen2 cannot
infer the implementations for dplyr to be S3 methods.
That in turn creates issues with CRAN runtime checks.

We can utilize the delayed registration mechanism in R-3.6 to achieve
compatibility. This means we will need to use `@rawNamespace` directives
instead of `@export`.

This is also a good time to bump the dplyr minimum version to a point
after the dbplyr split, since `src_desc` and `src_translate_env` have
been long gone. Trying to maintain compatibility with the versions of
dplyr where they existed will complicate the NAMESPACE file even
further.